@@ -36,6 +36,12 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
     z-index: 6000000;
 }
 
+.sf-toolbarreset * {
+    -webkit-box-sizing: content-box;
+    -moz-box-sizing: content-box;
+    box-sizing: content-box;
+}
+
 .sf-toolbarreset {
     position: fixed;
     background-color: #f7f7f7;
@@ -199,6 +205,7 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
 .sf-toolbar-block .sf-toolbar-icon > a,
 .sf-toolbar-block .sf-toolbar-icon > span {
     display: block;
+    font-weight: normal;
     text-decoration: none;
     margin: 0;
     padding: 5px 8px;
@@ -211,6 +218,10 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
 .sf-toolbar-block .sf-toolbar-icon > a:link
 .sf-toolbar-block .sf-toolbar-icon > a:hover {
     color: black !important;
+}
+
+.sf-toolbar-block .sf-toolbar-icon > a[href]:after {
+    content: \"\";
 }
 
 .sf-toolbar-block .sf-toolbar-icon img {
@@ -291,9 +302,9 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
 
 /***** Override the setting when the toolbar is on the top *****/
 ";
-        // line 275
-        if (($this->getContext($context, "position") == "top")) {
-            // line 276
+        // line 286
+        if (((isset($context["position"]) ? $context["position"] : $this->getContext($context, "position")) == "top")) {
+            // line 287
             echo "    .sf-minitoolbar {
         top: 0;
         bottom: auto;
@@ -335,24 +346,24 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
     }
 ";
         }
-        // line 316
+        // line 327
         echo "
 ";
-        // line 317
-        if ((!$this->getContext($context, "floatable"))) {
-            // line 318
+        // line 328
+        if ((!(isset($context["floatable"]) ? $context["floatable"] : $this->getContext($context, "floatable")))) {
+            // line 329
             echo "    .sf-toolbarreset {
         position: static;
         background: #cbcbcb;
 
-        background-image: -moz-linear-gradient(90deg, #cbcbcb, #e8e8e8); !important;
-        background-image: -webkit-gradient(linear, 0% 0%, 100% 0%, from(#cbcbcb), to(#e8e8e8)); !important;
-        background-image: -o-linear-gradient(180deg, #cbcbcb, #e8e8e8); !important;
-        background: linear-gradient(90deg, #cbcbcb, #e8e8e8); !important;
+        background-image: -moz-linear-gradient(90deg, #cbcbcb, #e8e8e8) !important;
+        background-image: -webkit-gradient(linear, 0% 0%, 100% 0%, from(#cbcbcb), to(#e8e8e8)) !important;
+        background-image: -o-linear-gradient(180deg, #cbcbcb, #e8e8e8) !important;
+        background: linear-gradient(90deg, #cbcbcb, #e8e8e8) !important;
     }
 ";
         }
-        // line 328
+        // line 339
         echo "
 /***** Media query *****/
 @media screen and (max-width: 779px) {
@@ -389,6 +400,14 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
         display: none;
     }
 }
+
+/***** Media query print: Do not print the Toolbar. *****/
+@media print {
+    .sf-toolbar {
+        display: none;
+        visibility: hidden;
+    }
+}
 ";
     }
 
@@ -404,6 +423,6 @@ class __TwigTemplate_16c317490d622d730d3d5142dfc5f210a68d15a08f6a645484b82e8b382
 
     public function getDebugInfo()
     {
-        return array (  356 => 328,  339 => 316,  295 => 275,  24 => 3,  19 => 1,  480 => 162,  474 => 161,  469 => 158,  461 => 155,  457 => 153,  444 => 149,  440 => 148,  437 => 147,  435 => 146,  430 => 144,  427 => 143,  423 => 142,  413 => 134,  409 => 132,  407 => 131,  402 => 130,  398 => 129,  393 => 126,  384 => 121,  381 => 120,  379 => 119,  374 => 116,  368 => 112,  365 => 111,  362 => 110,  360 => 109,  337 => 103,  322 => 101,  314 => 99,  305 => 95,  294 => 90,  285 => 89,  268 => 85,  264 => 84,  252 => 80,  247 => 78,  214 => 69,  177 => 65,  169 => 60,  132 => 51,  128 => 49,  107 => 36,  93 => 38,  273 => 96,  254 => 92,  240 => 86,  238 => 85,  230 => 82,  221 => 77,  219 => 76,  217 => 75,  204 => 72,  179 => 69,  171 => 61,  159 => 61,  138 => 54,  135 => 53,  78 => 21,  71 => 17,  209 => 82,  193 => 73,  149 => 51,  133 => 42,  103 => 32,  95 => 39,  86 => 24,  57 => 11,  48 => 8,  51 => 10,  34 => 5,  31 => 4,  806 => 488,  803 => 487,  792 => 485,  788 => 484,  784 => 482,  771 => 481,  745 => 476,  742 => 475,  723 => 473,  706 => 472,  702 => 470,  698 => 469,  694 => 468,  690 => 467,  686 => 466,  682 => 465,  678 => 464,  675 => 463,  673 => 462,  656 => 461,  645 => 460,  630 => 455,  625 => 453,  621 => 452,  618 => 451,  616 => 450,  602 => 449,  565 => 414,  547 => 411,  530 => 410,  527 => 409,  525 => 408,  520 => 406,  515 => 404,  244 => 136,  199 => 67,  196 => 92,  188 => 90,  182 => 66,  173 => 65,  68 => 30,  62 => 27,  28 => 3,  357 => 123,  344 => 318,  341 => 105,  332 => 116,  327 => 114,  324 => 113,  318 => 111,  306 => 107,  297 => 276,  291 => 102,  263 => 95,  258 => 81,  243 => 88,  231 => 83,  224 => 71,  212 => 78,  202 => 77,  190 => 76,  187 => 70,  174 => 65,  143 => 56,  136 => 71,  122 => 43,  117 => 39,  112 => 35,  104 => 32,  85 => 25,  75 => 19,  58 => 25,  44 => 10,  161 => 63,  158 => 80,  154 => 58,  151 => 59,  140 => 55,  125 => 44,  121 => 50,  118 => 49,  100 => 39,  87 => 25,  49 => 14,  46 => 12,  27 => 3,  91 => 27,  88 => 25,  63 => 15,  389 => 160,  386 => 159,  378 => 157,  371 => 156,  367 => 155,  363 => 126,  358 => 151,  353 => 121,  345 => 147,  343 => 146,  340 => 145,  334 => 141,  331 => 140,  328 => 139,  326 => 138,  321 => 112,  309 => 97,  307 => 128,  302 => 125,  296 => 121,  293 => 120,  290 => 119,  288 => 101,  283 => 88,  281 => 114,  276 => 111,  274 => 97,  269 => 94,  265 => 96,  259 => 103,  255 => 93,  253 => 100,  235 => 74,  232 => 88,  227 => 81,  222 => 83,  210 => 77,  208 => 68,  189 => 71,  184 => 63,  175 => 65,  170 => 84,  166 => 54,  163 => 62,  155 => 47,  152 => 46,  144 => 49,  127 => 35,  109 => 34,  94 => 28,  82 => 22,  76 => 28,  61 => 23,  39 => 6,  36 => 8,  79 => 21,  72 => 27,  69 => 26,  54 => 10,  47 => 9,  42 => 11,  40 => 11,  37 => 10,  22 => 2,  164 => 59,  157 => 56,  145 => 74,  139 => 49,  131 => 52,  120 => 31,  115 => 39,  111 => 37,  108 => 47,  106 => 33,  101 => 43,  98 => 31,  92 => 27,  83 => 35,  80 => 29,  74 => 14,  66 => 11,  60 => 13,  55 => 24,  52 => 12,  50 => 10,  41 => 19,  32 => 7,  29 => 3,  462 => 202,  453 => 151,  449 => 198,  446 => 197,  441 => 196,  439 => 195,  431 => 189,  429 => 188,  422 => 184,  415 => 180,  408 => 176,  401 => 172,  394 => 168,  387 => 122,  380 => 158,  373 => 156,  361 => 152,  355 => 106,  351 => 120,  348 => 140,  342 => 317,  338 => 135,  335 => 134,  329 => 131,  325 => 129,  323 => 128,  320 => 127,  315 => 110,  312 => 98,  303 => 106,  300 => 105,  298 => 91,  289 => 113,  286 => 112,  278 => 86,  275 => 105,  270 => 102,  267 => 101,  262 => 98,  256 => 96,  248 => 97,  246 => 90,  241 => 77,  233 => 87,  229 => 73,  226 => 84,  220 => 70,  216 => 79,  213 => 78,  207 => 75,  203 => 78,  200 => 72,  197 => 69,  194 => 68,  191 => 67,  185 => 74,  181 => 65,  178 => 66,  176 => 64,  172 => 64,  168 => 62,  165 => 83,  162 => 59,  156 => 62,  153 => 77,  150 => 55,  147 => 58,  141 => 48,  134 => 54,  130 => 41,  123 => 61,  119 => 42,  116 => 41,  113 => 48,  105 => 25,  102 => 32,  99 => 31,  96 => 37,  90 => 37,  84 => 40,  81 => 23,  73 => 19,  70 => 15,  67 => 15,  64 => 24,  59 => 22,  53 => 12,  45 => 8,  43 => 12,  38 => 6,  35 => 5,  33 => 4,  30 => 3,);
+        return array (  350 => 327,  308 => 287,  24 => 3,  19 => 1,  347 => 134,  319 => 124,  301 => 117,  299 => 116,  277 => 109,  271 => 108,  260 => 104,  257 => 103,  251 => 101,  239 => 97,  228 => 88,  225 => 87,  211 => 81,  148 => 60,  480 => 162,  474 => 161,  469 => 158,  461 => 155,  457 => 153,  444 => 149,  440 => 148,  437 => 147,  435 => 146,  430 => 144,  427 => 143,  423 => 142,  413 => 134,  409 => 132,  407 => 131,  402 => 130,  398 => 129,  393 => 126,  384 => 121,  381 => 120,  379 => 119,  374 => 116,  368 => 112,  365 => 141,  362 => 110,  360 => 109,  337 => 103,  322 => 101,  314 => 99,  305 => 119,  294 => 90,  285 => 111,  268 => 107,  264 => 84,  252 => 80,  247 => 78,  214 => 69,  177 => 65,  169 => 66,  140 => 55,  132 => 51,  128 => 49,  107 => 36,  93 => 38,  273 => 96,  254 => 102,  240 => 86,  238 => 85,  230 => 82,  221 => 77,  219 => 84,  217 => 75,  179 => 69,  159 => 61,  135 => 53,  78 => 21,  71 => 17,  209 => 82,  193 => 73,  154 => 58,  149 => 51,  133 => 42,  103 => 32,  95 => 39,  86 => 24,  57 => 11,  48 => 8,  51 => 10,  34 => 4,  31 => 3,  810 => 492,  807 => 491,  796 => 489,  792 => 488,  788 => 486,  775 => 485,  749 => 479,  746 => 478,  727 => 476,  710 => 475,  706 => 473,  702 => 472,  698 => 471,  694 => 470,  690 => 469,  686 => 468,  682 => 467,  679 => 466,  677 => 465,  660 => 464,  649 => 462,  634 => 456,  629 => 454,  625 => 453,  622 => 452,  620 => 451,  606 => 449,  601 => 446,  567 => 414,  549 => 411,  532 => 410,  529 => 409,  527 => 408,  522 => 406,  517 => 404,  199 => 67,  196 => 92,  182 => 69,  173 => 65,  158 => 62,  68 => 30,  62 => 27,  28 => 3,  357 => 123,  344 => 119,  341 => 131,  332 => 116,  327 => 126,  324 => 113,  318 => 111,  306 => 286,  297 => 104,  291 => 102,  263 => 95,  258 => 81,  243 => 98,  231 => 83,  224 => 71,  212 => 78,  202 => 77,  190 => 76,  187 => 70,  161 => 63,  143 => 59,  136 => 71,  122 => 43,  117 => 39,  112 => 35,  85 => 25,  75 => 19,  58 => 25,  44 => 10,  204 => 78,  188 => 90,  174 => 67,  171 => 61,  167 => 71,  138 => 57,  125 => 44,  121 => 50,  118 => 49,  104 => 32,  87 => 25,  49 => 9,  46 => 12,  27 => 3,  91 => 34,  88 => 25,  63 => 15,  389 => 160,  386 => 159,  378 => 157,  371 => 156,  367 => 339,  363 => 126,  358 => 139,  353 => 328,  345 => 147,  343 => 132,  340 => 145,  334 => 141,  331 => 140,  328 => 139,  326 => 138,  321 => 112,  309 => 120,  307 => 128,  302 => 125,  296 => 121,  293 => 113,  290 => 119,  288 => 101,  283 => 88,  281 => 110,  276 => 111,  274 => 97,  269 => 94,  265 => 106,  259 => 103,  255 => 93,  253 => 100,  235 => 74,  232 => 88,  227 => 81,  222 => 83,  210 => 77,  208 => 68,  189 => 71,  184 => 63,  175 => 58,  170 => 84,  166 => 54,  163 => 62,  155 => 47,  152 => 46,  144 => 49,  127 => 54,  109 => 34,  94 => 35,  82 => 28,  76 => 28,  61 => 23,  39 => 6,  36 => 8,  79 => 21,  72 => 27,  69 => 26,  54 => 10,  47 => 9,  42 => 11,  40 => 11,  37 => 10,  22 => 2,  164 => 59,  157 => 56,  145 => 74,  139 => 49,  131 => 55,  120 => 31,  115 => 39,  111 => 37,  108 => 47,  106 => 33,  101 => 43,  98 => 37,  92 => 27,  83 => 35,  80 => 29,  74 => 23,  66 => 11,  60 => 15,  55 => 24,  52 => 12,  50 => 10,  41 => 19,  32 => 7,  29 => 3,  462 => 202,  453 => 151,  449 => 198,  446 => 197,  441 => 196,  439 => 195,  431 => 189,  429 => 188,  422 => 184,  415 => 180,  408 => 176,  401 => 172,  394 => 168,  387 => 122,  380 => 158,  373 => 156,  361 => 152,  355 => 329,  351 => 135,  348 => 140,  342 => 137,  338 => 130,  335 => 134,  329 => 131,  325 => 129,  323 => 125,  320 => 127,  315 => 123,  312 => 98,  303 => 106,  300 => 105,  298 => 91,  289 => 112,  286 => 112,  278 => 86,  275 => 105,  270 => 102,  267 => 101,  262 => 105,  256 => 96,  248 => 100,  246 => 99,  241 => 77,  233 => 87,  229 => 73,  226 => 84,  220 => 70,  216 => 79,  213 => 82,  207 => 75,  203 => 78,  200 => 72,  197 => 74,  194 => 68,  191 => 77,  185 => 74,  181 => 65,  178 => 59,  176 => 64,  172 => 57,  168 => 62,  165 => 64,  162 => 59,  156 => 62,  153 => 77,  150 => 55,  147 => 58,  141 => 48,  134 => 56,  130 => 41,  123 => 61,  119 => 42,  116 => 45,  113 => 44,  105 => 39,  102 => 32,  99 => 31,  96 => 37,  90 => 37,  84 => 40,  81 => 23,  73 => 19,  70 => 15,  67 => 19,  64 => 24,  59 => 22,  53 => 10,  45 => 8,  43 => 12,  38 => 6,  35 => 5,  33 => 4,  30 => 3,);
     }
 }
