@@ -17,7 +17,7 @@ class __TwigTemplate_5160e554ef7cab9a3b34c8e219ed8e56a7aa0f3850abb1dede92722b749
     {
         // line 1
         echo "<div id=\"sfwdt";
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "\" class=\"sf-toolbar\" style=\"display: none\"></div>
 ";
         // line 2
@@ -27,10 +27,10 @@ class __TwigTemplate_5160e554ef7cab9a3b34c8e219ed8e56a7aa0f3850abb1dede92722b749
     (function () {
         ";
         // line 5
-        if (("top" == $this->getContext($context, "position"))) {
+        if (("top" == (isset($context["position"]) ? $context["position"] : $this->getContext($context, "position")))) {
             // line 6
             echo "            var sfwdt = document.getElementById('sfwdt";
-            echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+            echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
             echo "');
             document.body.insertBefore(
                 document.body.removeChild(sfwdt),
@@ -43,11 +43,11 @@ class __TwigTemplate_5160e554ef7cab9a3b34c8e219ed8e56a7aa0f3850abb1dede92722b749
         Sfjs.load(
             'sfwdt";
         // line 14
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "',
             '";
         // line 15
-        echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("_wdt", array("token" => $this->getContext($context, "token"))), "html", null, true);
+        echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("_wdt", array("token" => (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")))), "html", null, true);
         echo "',
             function(xhr, el) {
                 el.style.display = -1 !== xhr.responseText.indexOf('sf-toolbarreset') ? 'block' : 'none';
@@ -59,28 +59,28 @@ class __TwigTemplate_5160e554ef7cab9a3b34c8e219ed8e56a7aa0f3850abb1dede92722b749
                 if (Sfjs.getPreference('toolbar/displayState') == 'none') {
                     document.getElementById('sfToolbarMainContent-";
         // line 24
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "').style.display = 'none';
                     document.getElementById('sfToolbarClearer-";
         // line 25
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "').style.display = 'none';
                     document.getElementById('sfMiniToolbar-";
         // line 26
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "').style.display = 'block';
                 } else {
                     document.getElementById('sfToolbarMainContent-";
         // line 28
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "').style.display = 'block';
                     document.getElementById('sfToolbarClearer-";
         // line 29
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "').style.display = 'block';
                     document.getElementById('sfMiniToolbar-";
         // line 30
-        echo twig_escape_filter($this->env, $this->getContext($context, "token"), "html", null, true);
+        echo twig_escape_filter($this->env, (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")), "html", null, true);
         echo "').style.display = 'none';
                 }
             },
@@ -88,10 +88,11 @@ class __TwigTemplate_5160e554ef7cab9a3b34c8e219ed8e56a7aa0f3850abb1dede92722b749
                 if (xhr.status !== 0) {
                     confirm('An error occurred while loading the web debug toolbar (' + xhr.status + ': ' + xhr.statusText + ').\\n\\nDo you want to open the profiler?') && (window.location = '";
         // line 35
-        echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("_profiler", array("token" => $this->getContext($context, "token"))), "html", null, true);
+        echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("_profiler", array("token" => (isset($context["token"]) ? $context["token"] : $this->getContext($context, "token")))), "html", null, true);
         echo "');
                 }
-            }
+            },
+            {'maxTries': 5}
         );
     })();
 /*]]>*/</script>
@@ -110,6 +111,6 @@ class __TwigTemplate_5160e554ef7cab9a3b34c8e219ed8e56a7aa0f3850abb1dede92722b749
 
     public function getDebugInfo()
     {
-        return array (  91 => 35,  83 => 30,  79 => 29,  75 => 28,  70 => 26,  66 => 25,  62 => 24,  50 => 15,  46 => 14,  42 => 12,  32 => 6,  30 => 5,  26 => 3,  24 => 2,  19 => 1,  729 => 15,  722 => 10,  717 => 7,  714 => 6,  689 => 480,  682 => 476,  652 => 449,  630 => 430,  621 => 424,  610 => 416,  603 => 412,  595 => 407,  587 => 402,  570 => 388,  560 => 381,  554 => 378,  546 => 373,  532 => 362,  524 => 357,  506 => 342,  496 => 335,  479 => 321,  465 => 310,  455 => 303,  448 => 299,  440 => 294,  432 => 289,  422 => 282,  414 => 277,  399 => 265,  383 => 252,  373 => 245,  367 => 242,  361 => 239,  354 => 235,  346 => 230,  338 => 225,  331 => 221,  320 => 213,  309 => 205,  301 => 200,  285 => 187,  270 => 175,  261 => 169,  254 => 165,  245 => 159,  227 => 144,  219 => 139,  209 => 132,  199 => 125,  192 => 121,  185 => 117,  177 => 112,  150 => 88,  144 => 85,  139 => 83,  134 => 81,  123 => 73,  116 => 69,  93 => 57,  84 => 51,  78 => 48,  63 => 36,  55 => 31,  47 => 26,  35 => 16,  33 => 15,  29 => 13,  27 => 6,  21 => 2,);
+        return array (  91 => 35,  83 => 30,  79 => 29,  75 => 28,  70 => 26,  66 => 25,  62 => 24,  50 => 15,  46 => 14,  42 => 12,  32 => 6,  30 => 5,  26 => 3,  24 => 2,  19 => 1,  724 => 15,  717 => 10,  712 => 7,  709 => 6,  684 => 475,  677 => 471,  647 => 444,  625 => 425,  616 => 419,  605 => 411,  598 => 407,  590 => 402,  582 => 397,  565 => 383,  555 => 376,  549 => 373,  541 => 368,  527 => 357,  519 => 352,  501 => 337,  491 => 330,  474 => 316,  460 => 305,  450 => 298,  443 => 294,  435 => 289,  427 => 284,  417 => 277,  409 => 272,  394 => 260,  378 => 247,  368 => 240,  362 => 237,  355 => 233,  347 => 228,  339 => 223,  331 => 218,  324 => 214,  313 => 206,  302 => 198,  294 => 193,  278 => 180,  263 => 168,  254 => 162,  247 => 158,  238 => 152,  220 => 137,  212 => 132,  202 => 125,  192 => 118,  185 => 114,  178 => 110,  170 => 105,  143 => 81,  137 => 78,  133 => 77,  129 => 76,  120 => 70,  115 => 68,  93 => 57,  84 => 51,  78 => 48,  63 => 36,  55 => 31,  47 => 26,  35 => 16,  33 => 15,  29 => 13,  27 => 6,  21 => 2,);
     }
 }
