@@ -27,12 +27,26 @@ class __TwigTemplate_66b9528a08be348045cb6590e3aeb61f8870b74152ad5755f1356d72275
             request = function(url, onSuccess, onError, payload, options) {
                 var xhr = window.XMLHttpRequest ? new XMLHttpRequest() : new ActiveXObject('Microsoft.XMLHTTP');
                 options = options || {};
+                options.maxTries = options.maxTries || 0;
                 xhr.open(options.method || 'GET', url, true);
                 xhr.setRequestHeader('X-Requested-With', 'XMLHttpRequest');
                 xhr.onreadystatechange = function(state) {
-                    if (4 === xhr.readyState && 200 === xhr.status) {
+                    if (4 !== xhr.readyState) {
+                        return null;
+                    }
+
+                    if (xhr.status == 404 && options.maxTries > 1) {
+                        setTimeout(function(){
+                            options.maxTries--;
+                            request(url, onSuccess, onError, payload, options);
+                        }, 500);
+
+                        return null;
+                    }
+
+                    if (200 === xhr.status) {
                         (onSuccess || noop)(xhr);
-                    } else if (4 === xhr.readyState && xhr.status != 200) {
+                    } else {
                         (onError || noop)(xhr);
                     }
                 };
@@ -93,6 +107,7 @@ class __TwigTemplate_66b9528a08be348045cb6590e3aeb61f8870b74152ad5755f1356d72275
                             (onSuccess || noop)(xhr, el);
                         },
                         function(xhr) { (onError || noop)(xhr, el); },
+                        '',
                         options
                     );
                 }
@@ -128,6 +143,6 @@ class __TwigTemplate_66b9528a08be348045cb6590e3aeb61f8870b74152ad5755f1356d72275
 
     public function getDebugInfo()
     {
-        return array (  91 => 35,  83 => 30,  79 => 29,  75 => 28,  70 => 26,  66 => 25,  62 => 24,  50 => 15,  46 => 14,  42 => 12,  32 => 6,  30 => 5,  26 => 3,  24 => 2,  19 => 1,  729 => 15,  722 => 10,  717 => 7,  714 => 6,  689 => 480,  682 => 476,  652 => 449,  630 => 430,  621 => 424,  610 => 416,  603 => 412,  595 => 407,  587 => 402,  570 => 388,  560 => 381,  554 => 378,  546 => 373,  532 => 362,  524 => 357,  506 => 342,  496 => 335,  479 => 321,  465 => 310,  455 => 303,  448 => 299,  440 => 294,  432 => 289,  422 => 282,  414 => 277,  399 => 265,  383 => 252,  373 => 245,  367 => 242,  361 => 239,  354 => 235,  346 => 230,  338 => 225,  331 => 221,  320 => 213,  309 => 205,  301 => 200,  285 => 187,  270 => 175,  261 => 169,  254 => 165,  245 => 159,  227 => 144,  219 => 139,  209 => 132,  199 => 125,  192 => 121,  185 => 117,  177 => 112,  150 => 88,  144 => 85,  139 => 83,  134 => 81,  123 => 73,  116 => 69,  93 => 57,  84 => 51,  78 => 48,  63 => 36,  55 => 31,  47 => 26,  35 => 16,  33 => 15,  29 => 13,  27 => 6,  21 => 2,);
+        return array (  91 => 35,  83 => 30,  79 => 29,  75 => 28,  70 => 26,  66 => 25,  62 => 24,  50 => 15,  46 => 14,  42 => 12,  32 => 6,  30 => 5,  26 => 3,  24 => 2,  19 => 1,  724 => 15,  717 => 10,  712 => 7,  709 => 6,  684 => 475,  677 => 471,  647 => 444,  625 => 425,  616 => 419,  605 => 411,  598 => 407,  590 => 402,  582 => 397,  565 => 383,  555 => 376,  549 => 373,  541 => 368,  527 => 357,  519 => 352,  501 => 337,  491 => 330,  474 => 316,  460 => 305,  450 => 298,  443 => 294,  435 => 289,  427 => 284,  417 => 277,  409 => 272,  394 => 260,  378 => 247,  368 => 240,  362 => 237,  355 => 233,  347 => 228,  339 => 223,  331 => 218,  324 => 214,  313 => 206,  302 => 198,  294 => 193,  278 => 180,  263 => 168,  254 => 162,  247 => 158,  238 => 152,  220 => 137,  212 => 132,  202 => 125,  192 => 118,  185 => 114,  178 => 110,  170 => 105,  143 => 81,  137 => 78,  133 => 77,  129 => 76,  120 => 70,  115 => 68,  93 => 57,  84 => 51,  78 => 48,  63 => 36,  55 => 31,  47 => 26,  35 => 16,  33 => 15,  29 => 13,  27 => 6,  21 => 2,);
     }
 }
